chore(create-post-page): separating front-end logic on create post page
