+++ src/app/post/create/page.tsx
@@ -1,29 +1,21 @@
 "use client";
 
 import dynamic from "next/dynamic";
-import { useState } from "react";
 import { useCreatePost } from "./page.client";
 
 const MDEditor = dynamic(() => import("@uiw/react-md-editor"), { ssr: false });
 
 const Page = () => {
-  const { createPost, isUploading, successMessage, errorMessage } =
-    useCreatePost();
-
-  const [title, setTitle] = useState<string>("");
-  const [content, setContent] = useState<string>("");
-
-  const handleCreatePost = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    if (!title || !content) {
-      return alert("Title and content are required.");
-    }
-
-    createPost({
-      title,
-      content
-    });
-  };
+  const {
+    title,
+    setTitle,
+    content,
+    setContent,
+    handleCreatePost,
+    isUploading,
+    successMessage,
+    errorMessage
+  } = useCreatePost();
 
   return (
     <div className="flex flex-col items-center justify-center gap-4 py-8 md:py-10">
